requirements.txt
requests==2.32.3
fastapi==0.111.0
fastapi_cors==0.0.6
uvicorn==0.30.1
# annotated-types==0.7.0
# anyio==4.4.0
# certifi==2024.6.2
# click==8.1.7
# dnspython==2.6.1
# email_validator==2.2.0
# environs==11.0.0
# exceptiongroup==1.2.1
# fastapi-cli==0.0.4
# h11==0.14.0
# httpcore==1.0.5
# httptools==0.6.1
# httpx==0.27.0
# idna==3.7
# Jinja2==3.1.4
# markdown-it-py==3.0.0
# MarkupSafe==2.1.5
# marshmallow==3.21.3
# mdurl==0.1.2
# orjson==3.10.5
# packaging==24.1
#pydantic==2.8.2
#pydantic_core==2.20.1
#pydantic-settings     2.3.4

#pydantic==2.7.4
#pydantic_core==2.18.4
# Pygments==2.18.0
python-dotenv==1.0.1
python-multipart==0.0.9
# PyYAML==6.0.1
# rich==13.7.1
# shellingham==1.5.4
# sniffio==1.3.1
# starlette==0.37.2
# typer==0.12.3
# typing_extensions==4.12.2
# ujson==5.10.0
# uvloop==0.19.0
# watchfiles==0.22.0
# websockets==12.0
phidata==2.4.25
openai==1.35.10
duckduckgo_search==6.1.9
yfinance==0.2.40